Move siu parser to own file
